bgzf: Work with append mode
Previously, the BgzfReader would assume that the file ended when it
encountered an EOF marker. However, this would not play well with files
that were written to, closed, re-opened in append mode, and had further
data written -- the appended data would be inaccessible.

Note that a similar situation would arise if you called flush() on a
BgzfWriter twice without writing anything in between: the second flush
would go ahead and write a block to disk for the empty buffer, which is
equivalent to an EOF marker.

Now, keep reading until we actually reach the end of the file, as
indicated by a block_size of zero.

Fixes #2142